feat: AppRouter, TitleUpdater, modif App.js
