--- frontend/public/scripts-backup
@@ -1,71 +1,0 @@
-        <!-- АКТУАЛЬНОЕ !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!-->
-<!--    <script>-->
-<!--        fetch('http://localhost:8000/api/information')-->
-<!--        .then(response => response.json())-->
-<!--        .then(data => {-->
-<!--            const tableBody = document.querySelector('#datatablesSimple tbody');-->
-<!--            tableBody.innerHTML = '';-->
-
-<!--            data.forEach(item => {-->
-<!--                const row = document.createElement('tr');-->
-<!--                row.innerHTML = `-->
-<!--                    <td>${item.org_name}</td>-->
-<!--                    <td>${item.deep}</td>-->
-<!--                    <td>${item.average}</td>-->
-<!--                    <td>${item.active}</td>-->
-<!--                    `;-->
-<!--                tableBody.appendChild(row);-->
-<!--            });-->
-<!--        })-->
-<!--        .catch(error => console.error('Errors:', error));-->
-<!--    </script>-->
-
-<!--    <script>-->
-<!--        fetch('http://localhost:8000/api/information')-->
-<!--        .then(response => response.json())-->
-<!--        .then(data => {-->
-<!--            const tableBody = document.querySelector('#datatablesSimple tbody');-->
-<!--            tableBody.innerHTML = '';-->
-
-<!--            data.forEach(item => {-->
-<!--                row = document.createElement('tr');-->
-<!--                row.innerHTML = `-->
-<!--                    <td>${item.org_name}</td>-->
-<!--                    <td>${item.deep}</td>-->
-<!--                    <td>${item.average}</td>-->
-<!--                    <td>${item.active}</td>-->
-<!--                `;-->
-<!--                tableBody.appendChild(row);-->
-<!--            });-->
-<!--        })-->
-<!--        .catch(error => console.error('Errors: ', error));-->
-<!--    </script>-->
-
-
-<!--    <script>-->
-<!--        fetch('http://localhost:8000/test/').then(response => response.json()).then(data => console.log(data));-->
-<!--    </script>-->
-
-
-
-<!-- СТАРОЕ-->
-<!--    <script>-->
-<!--          fetch('http://localhost:8000/api/information/')-->
-<!--          .then(response => response.json())-->
-<!--          .then(data => {-->
-<!--                const tableBody = document.querySelector('#datatablesSimple tbody');-->
-<!--                tableBody.innerHTML = ''; // Очистка таблицы перед добавлением новых данных-->
-
-<!--                data.forEach(item => {-->
-<!--                    const row = document.createElement('tr');-->
-<!--                    row.innerHTML = `-->
-<!--                        <td>${item.org_name}</td>-->
-<!--                        <td>${item.deep}</td>-->
-<!--                        <td>${item.average}</td>-->
-<!--                        <td>${item.active}</td>-->
-<!--                    `;-->
-<!--                    tableBody.appendChild(row);-->
-<!--                });-->
-<!--          })-->
-<!--          .catch(error => console.error('Error:', error));-->
-<!--    </script>-->
