+++ frontend/src/App.js
@@ -1,41 +1,12 @@
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import { useState } from 'react';
-
-import Header from './components/Header/Header';
-import Sidebar from './components/Sidebar/Sidebar';
-// import Home from './components/Home/Home';
-import Main from './components/Main/Main';
-import About from './components/About/About';
-import Navbar from './components/Navbar/Navbar';
+import React from 'react';
+import AppRouter from './components/AppRouter/AppRouter';
 
 function App() {
 
-    // Бургер-меню
-    const [isSidebarToggled, setIsSidebarToggled] = useState(false);
-    // Переключание состояния сайтбара
-    const toggleSidebar = () => {
-        setIsSidebarToggled(prevState => !prevState); // Обновляем состояние, инвертируя предыдущее значение
-        localStorage.setItem('sb|sidebar-toggle', !isSidebarToggled); // Сохраняем текущее состояние в localStorage
-      };
-
   return (
-    <Router>
-      <div className={`App ${isSidebarToggled ? 'sb-sidenav-toggled' : ''}`}>
-        <Navbar />
-        <Header toggleSidebar={toggleSidebar} />
-
-        <Routes>
-            <Route path="/" element={<Main />} />
-            <Route path="/about-page" element={<About />} />
-        </Routes>
-
-        <div id="layoutSidenav">
-            <Sidebar />
-
-        </div>
-
-      </div>
-    </Router>
+     <div className="App">
+        <AppRouter />
+     </div>
   );
 }
